Renamed app to c2f
