+++ backend/c2g/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class Census2FacConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "c2g"
