--- backend/census2fac/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class Census2FacConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "census2fac"
